Refactor cli package.
- decouple __main__.py and core.py to solve import issues.
