+++ aea/cli/__main__.py
@@ -20,116 +20,14 @@
 
 """Entry-point for the AEA command-line tool."""
 
-import os
-import shutil
-from pathlib import Path
-from typing import cast
-
-import click
-from click import pass_context
-from jsonschema import ValidationError
-
-import aea
-from aea.cli.add import connection, add, skill
-from aea.cli.common import Context, pass_ctx, logger, _try_to_load_agent_config, DEFAULT_REGISTRY_PATH
+from aea.cli.add import add
 from aea.cli.install import install
 from aea.cli.list import list as _list
-from aea.cli.loggers import simple_verbosity_option
 from aea.cli.remove import remove
 from aea.cli.run import run
 from aea.cli.scaffold import scaffold
 from aea.cli.search import search
-from aea.configurations.base import DEFAULT_AEA_CONFIG_FILE, AgentConfig
-
-DEFAULT_CONNECTION = "oef"
-DEFAULT_SKILL = "error"
-
-
-@click.group()
-@click.version_option(aea.__version__)
-@simple_verbosity_option(logger, default="INFO")
-@click.pass_context
-def cli(ctx) -> None:
-    """Command-line tool for setting up an Autonomous Economic Agent."""
-    ctx.obj = Context(cwd=".")
-
-
-@cli.command()
-@click.argument('agent_name', type=str, required=True)
-@pass_context
-def create(click_context, agent_name):
-    """Create an agent."""
-    ctx = cast(Context, click_context.obj)
-    path = Path(agent_name)
-    logger.info("Creating agent's directory in '{}'".format(path))
-
-    # create the agent's directory
-    try:
-        path.mkdir(exist_ok=False)
-
-        # create a config file inside it
-        config_file = open(os.path.join(agent_name, DEFAULT_AEA_CONFIG_FILE), "w")
-        agent_config = AgentConfig(agent_name=agent_name, aea_version=aea.__version__, authors="", version="v1", license="", url="", registry_path=DEFAULT_REGISTRY_PATH, description="")
-        agent_config.default_connection = DEFAULT_CONNECTION
-        ctx.agent_loader.dump(agent_config, config_file)
-        logger.info("Created config file {}".format(DEFAULT_AEA_CONFIG_FILE))
-
-        # next commands must be done from the agent's directory -> overwrite ctx.cwd
-        ctx.agent_config = agent_config
-        ctx.cwd = agent_config.agent_name
-
-        logger.info("Adding default connection '{}' to the agent...".format(DEFAULT_CONNECTION))
-        click_context.invoke(connection, connection_name=DEFAULT_CONNECTION)
-
-        logger.info("Adding default skill '{}' to the agent...".format(DEFAULT_SKILL))
-        click_context.invoke(skill, skill_name=DEFAULT_SKILL)
-
-    except OSError:
-        logger.error("Directory already exist. Aborting...")
-        exit(-1)
-    except ValidationError as e:
-        logger.error(str(e))
-        shutil.rmtree(agent_name, ignore_errors=True)
-        exit(-1)
-    except Exception as e:
-        logger.exception(e)
-        shutil.rmtree(agent_name, ignore_errors=True)
-        exit(-1)
-
-
-@cli.command()
-@click.argument('agent_name', type=str, required=True)
-@pass_ctx
-def delete(ctx: Context, agent_name):
-    """Delete an agent."""
-    path = Path(agent_name)
-    logger.info("Deleting agent's directory in '{}'...".format(path))
-
-    # delete the agent's directory
-    try:
-        shutil.rmtree(path, ignore_errors=False)
-    except OSError:
-        logger.error("An error occurred while deleting the agent directory. Aborting...")
-        exit(-1)
-
-
-@cli.command()
-@pass_ctx
-def freeze(ctx: Context):
-    """Get the dependencies."""
-    _try_to_load_agent_config(ctx)
-    for d in ctx.get_dependencies():
-        print(d)
-
-
-@cli.command()
-@pass_ctx
-def gui(ctx: Context):
-    """Run the CLI GUI."""
-    import aea.cli_gui  # pragma: no cover
-    logger.info("Running the GUI.....(press Ctrl+C to exit)")   # pragma: no cover
-    aea.cli_gui.run()   # pragma: no cover
-
+from aea.cli.core import cli
 
 cli.add_command(add)
 cli.add_command(_list)
